requirements.txt
Flask>=2.3.0
numpy>=1.21.0
Pillow>=8.0.0
matplotlib>=3.5.0
noise>=1.2.0
scipy>=1.7.0